refactor env
